requirements.txt
Flask==3.1.0
openai>=2.6.0
python-dotenv==1.0.1
flask-sqlalchemy
gunicorn